requirements.txt
torch>=2.2,<3
gymnasium>=0.29,<1
pettingzoo>=1.24,<2
numpy>=1.26,<3
pandas>=2,<3
matplotlib>=3.7,<4
PyYAML>=6,<7
tqdm>=4.65,<5
imageio[ffmpeg]>=2.31,<3
pygame>=2.5,<3
